% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{bn24}
\alias{bn24}
\title{bn24: List of species that are included in the BirdNET 2.4 model}
\format{
Dataframe with rows and columns
\describe{
\item{birdnet_code}{Scientific name and common name of a species, joined by a "_"}
\item{SCINAME}{Scientific name of a bird species}
\item{COMMONNAME}{Common name of a bird species}}
@source https://github.com/kahst/BirdNET-Analyzer/
}
\usage{
bn24
}
\description{
bn24: List of species that are included in the BirdNET 2.4 model
}
\keyword{datasets}
