% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eblist.R
\name{get_eblist}
\alias{get_eblist}
\title{Scrape eBird species from hotspot}
\usage{
get_eblist(hotspot_id, out_name = hotspot_id, out_dir = getwd())
}
\arguments{
\item{hotspot_id}{the ID number given by eBird to a given hotspot or region}

\item{out_name}{the middle portion of the output filename, see example}

\item{out_dir}{the location where your list will export to. Defaults to current working directory}
}
\value{
Nothing, only an exported CSV
}
\description{
Scrape eBird species from hotspot
}
\examples{
# Export the species list from Surprise Canyon, CA
# URL for this hotspot is here: https://ebird.org/hotspot/L11704882
# use the bit after the last /
get_eblist("L11704882")
}
