% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bn_speclist.R
\name{make_bn_speclist}
\alias{make_bn_speclist}
\title{Load a custom birdlist and convert it to BirdNET species list format}
\usage{
make_bn_speclist(
  ebird_file,
  bn_version = "2.4",
  output_file,
  name_col = NULL,
  sci_col = NULL,
  del_ebird_file = F,
  ebird_dir = getwd(),
  out_dir = getwd(),
  test_run = F,
  add_species = NULL
)
}
\arguments{
\item{ebird_file}{The custom file you want to read in, must have COMMONNAME or SCINAME column}

\item{bn_version}{The version of BirdNET you plan to use, "2.1", "2.2", "2.4"}

\item{output_file}{The name of your output file, must end in ".txt"}

\item{name_col}{Specify the name of your column that holds common names or choose SCINAME}

\item{sci_col}{Specifiy the name of your column that holds scientific names or chose COMMONAME}

\item{del_ebird_file}{Set to TRUE if you want to delete your source species list on completion}

\item{ebird_dir}{The directory your ebird list is comiing from, defaults to current dir}

\item{out_dir}{The directory your BirdNET output will be written to, defaults to current dir}

\item{test_run}{Set to TRUE if you do not want an output file}

\item{add_species}{Supply a list of bird names that are missing from your custom list}
}
\value{
A table of birds that did not match the master BirdNET species list
}
\description{
Load a custom birdlist and convert it to BirdNET species list format
}
\examples{
missing_table <- make_bn_speclist(ebird_file = "ebird_list.csv", output = "birdnet_list.csv", name_col = "COMMONNAME", test_run = T)
}
